change ServerResponse to be BaseResponse & add SerialName & moved user mapper to new file
