+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/model/TaxiFiltrationDto.kt
@@ -1,7 +1,12 @@
 package org.thechance.common.data.remote.model
 
+import kotlinx.serialization.SerialName
+
 data class TaxiFiltrationDto(
+    @SerialName("color")
     val color: Int,
+    @SerialName("seats")
     val seats: Int,
+    @SerialName("status")
     val status: Int,
 )
